processing-worker/requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Machine Learning / NLP
numpy==1.26.4
scikit-learn==1.4.1.post1
spacy==3.7.4

# Clustering Dependencies
sentence-transformers==2.3.1
torch>=2.2.0  # User may install torch with XPU support separately
datasketch==1.6.4

# Utilities
psutil==5.9.8

# Testing
pytest==8.1.1
pytest-cov==4.1.0
